@@ -1,46 +1,51 @@
 {
   "name": "aspose_front",
   "description": "Front",
+  "license": "MIT",
+  "repository": {
+    "url": "https://github.com/Marat-Gumerov/aspo"
+  },
+  "readme": "../../README.md",
   "scripts": {
     "start": "ng serve",
     "build": "ng build"
   },
   "dependencies": {
-    "@angular/common": "^8.2.5",
-    "@angular/compiler": "^8.2.5",
-    "@angular/core": "^8.2.5",
-    "@angular/forms": "^8.2.5",
-    "@angular/platform-browser": "^8.2.5",
-    "@angular/platform-browser-dynamic": "^8.2.5",
-    "@angular/router": "^8.2.5",
-    "@angular/animations": "^8.2.5",
-    "@angular/material": "^8.1.4",
-    "@angular/cdk": "^8.1.4",
-    "bootstrap": "^4.3.1",
-    "core-js": "^3.2.1",
-    "font-awesome": "^4.7.0",
-    "jquery": "^3.4.1",
-    "rxjs": "^6.5.3",
-    "zone.js": "~0.9.1",
-    "popper.js": "1.14.7"
+    "@angular/common": "9.1.6",
+    "@angular/compiler": "9.1.6",
+    "@angular/core": "9.1.6",
+    "@angular/forms": "9.1.6",
+    "@angular/platform-browser": "9.1.6",
+    "@angular/platform-browser-dynamic": "9.1.6",
+    "@angular/router": "9.1.6",
+    "@angular/animations": "9.1.6",
+    "@angular/material": "9.2.3",
+    "@angular/cdk": "9.2.3",
+    "bootstrap": "4.4.1",
+    "core-js": "3.6.5",
+    "font-awesome": "4.7.0",
+    "jquery": "3.5.1",
+    "rxjs": "6.5.5",
+    "zone.js": "0.10.3",
+    "popper.js": "1.16.1"
   },
   "devDependencies": {
-    "@angular/cli": "^8.3.3",
-    "@angular-devkit/build-angular": "^0.803.3",
-    "@angular/compiler-cli": "^8.2.5",
-    "@types/node": "12.7.4",
-    "typescript": "3.5.3",
-    "webpack": "4.39.3",
-    "webpack-cli": "3.3.8",
-    "webpack-dev-server": "3.8.0",
+    "@angular/cli": "9.1.5",
+    "@angular-devkit/build-angular": "0.901.5",
+    "@angular/compiler-cli": "9.1.6",
+    "@types/node": "13.13.5",
+    "typescript": "3.8.3",
+    "webpack": "4.43.0",
+    "webpack-cli": "3.3.11",
+    "webpack-dev-server": "3.11.0",
     "angular2-template-loader": "0.6.2",
     "awesome-typescript-loader": "5.2.1",
     "uglifyjs-webpack-plugin": "2.2.0",
-    "raw-loader": "3.1.0",
-    "html-loader": "0.5.5",
-    "html-webpack-plugin": "3.2.0",
-    "css-loader": "3.2.0",
-    "file-loader": "4.2.0",
-    "mini-css-extract-plugin": "0.8.0"
+    "raw-loader": "4.0.1",
+    "html-loader": "1.1.0",
+    "html-webpack-plugin": "4.3.0",
+    "css-loader": "3.5.3",
+    "file-loader": "6.0.0",
+    "mini-css-extract-plugin": "0.9.0"
   }
 }
