@@ -1,10 +1,10 @@
-import { Inject, Component } from "@angular/core";
-import { MAT_DIALOG_DATA, MatDialogRef } from "@angular/material";
+import {Component, Inject} from "@angular/core";
+import {MAT_DIALOG_DATA, MatDialogRef} from "@angular/material/dialog";
 import Worker from "../../models/worker";
 import Util from "../../util/util";
-import { WorkerService } from "../../services/worker.service";
-import { Observable } from "rxjs";
-import { WorkerTypeService } from "../../services/worker.type.service";
+import {WorkerService} from "../../services/worker.service";
+import {Observable} from "rxjs";
+import {WorkerTypeService} from "../../services/worker.type.service";
 import WorkerType from "../../models/worker.type";
 
 @Component({
@@ -15,6 +15,7 @@ export class EditWorkerDialogComponent {
     state: string;
     workerTypes: WorkerType[];
     chiefs: Worker[];
+
     constructor(
         public dialogRef: MatDialogRef<EditWorkerDialogComponent>,
         @Inject(MAT_DIALOG_DATA) public worker: Worker,
@@ -38,8 +39,8 @@ export class EditWorkerDialogComponent {
         this.getChiefs();
     }
 
-    onSalaryBaseInput(value: number): void {
-        this.worker.salaryBase = value;
+    onSalaryBaseInput(value: string): void {
+        this.worker.salaryBase = Number(value);
     }
 
     onSaveClick(): void {
@@ -51,7 +52,7 @@ export class EditWorkerDialogComponent {
         }
         observable.subscribe({
             next: (worker) => {
-                this.worker = worker;
+                this.worker = Worker.fromData(worker);
                 this.dialogRef.close(worker);
             },
             error: (error) => console.error(error)
@@ -74,8 +75,10 @@ export class EditWorkerDialogComponent {
         this.workerTypeService.get()
             .subscribe({
                 next: (workerTypes) => {
-                    this.workerTypes = workerTypes;
-                    this.worker.workerType = this.worker.workerType ? undefined : this.workerTypes[0];
+                    this.workerTypes = WorkerType.fromDataList(workerTypes);
+                    this.worker.workerType = this.worker.workerType
+                        ? this.workerTypes.find(workerType => workerType.id === this.worker.workerType.id)
+                        : this.workerTypes[0];
                 },
                 error: (error) => console.error(error)
             });
@@ -85,7 +88,7 @@ export class EditWorkerDialogComponent {
         this.workerService.get(this.worker.employmentDate)
             .subscribe({
                 next: (workers) => {
-                    this.chiefs = workers;
+                    this.chiefs = Worker.fromDataList(workers);
                     this.chiefs.unshift(undefined);
                 },
                 error: (error) => console.error(error)
